requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.7
openai==1.3.8